1) Enchansed tag creation. 2) moved forms classes to a seperate forms.py file 3) deleted empty index.html
