+++ personal_assistant/notes/tests.py
@@ -1,4 +1,3 @@
-import pytest
 from django.test import Client, TestCase
 from django.urls import reverse
 from django.contrib.auth.models import User
